# arion/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
  handler.ErrorApiResponse:
    properties:
      code:
        type: string
      data:
        type: object
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handler.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  response.ProductData:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.UserData:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: |-
    Order management API for Jastipers

    All responses use this envelope:
    { "success": bool, "data": <payload>, "code": string, "message": string }
    Success: success=true, data holds the payload. Error: success=false, data={}, code and message describe the error.
  title: Recapo API
  version: "1.0"
paths:
...
  /product:
    post:
      consumes:
      - application/json
      description: |-
        Create a new product for the shop.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductData'
        "400":
          description: Bad request (invalid JSON or validation)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - product
  /products:
    get:
      consumes:
      - application/json
      description: |-
        Get all products for the shop. Optional search query to filter by name or description.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ProductData'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - product
  /products/{product_id}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete a product by ID.
        Success Response envelope: { success, data, code, message }. data contains "OK" on success.
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success. data contains \"OK\
          schema:
            type: string
        "400":
          description: Bad request (invalid product_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - product
    get:
      consumes:
      - application/json
      description: |-
        Get a single product by ID.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductData'
        "400":
          description: Bad request (invalid product_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - product
...
  /user:
    patch:
      consumes:
      - application/json
      description: |-
        Update the authenticated user's profile. Only provided fields are updated.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserData'
        "400":
          description: Bad request (invalid JSON)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      description: |-
        Get all users. Requires appropriate permissions.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.UserData'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - user
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"